refact: dry getBalance query
