+++ server/graphql/loaders/index.js
@@ -6,6 +6,7 @@ import { types as CollectiveType } from '../../constants/collectives';
 import { maxInteger } from '../../constants/math';
 import { TransactionTypes } from '../../constants/transactions';
 import { getListOfAccessibleMembers } from '../../lib/auth';
+import queries from '../../lib/queries';
 import models, { Op, sequelize } from '../../models';
 
 import collectiveLoaders from './collective';
@@ -54,38 +55,8 @@ export const loaders = req => {
 
   // Collective - Balance
   context.loaders.Collective.balance = new DataLoader(ids =>
-    sequelize
-      .query(
-        `
-        WITH "blockedFunds" AS (
-          SELECT
-            e."CollectiveId", COALESCE(sum(e.amount), 0) as sum
-          FROM
-            "Expenses" e
-          WHERE
-            e."CollectiveId" IN (:ids)
-            AND (
-              e.status = 'SCHEDULED_FOR_PAYMENT'
-              OR (
-                e.status = 'PROCESSING' AND e.data ->> 'payout_batch_id' IS NOT NULL
-              )
-          )
-          GROUP BY
-            e."CollectiveId"
-        )
-        SELECT
-          t."CollectiveId",
-          COALESCE(sum(t."netAmountInCollectiveCurrency") - COALESCE(max(bf.sum), 0), 0) AS "balance"
-        FROM
-          "Transactions" t
-        LEFT JOIN "blockedFunds" bf ON t."CollectiveId" = bf."CollectiveId"
-        WHERE
-          t."CollectiveId" IN (:ids)
-        GROUP BY
-          t."CollectiveId";
-      `,
-        { type: sequelize.QueryTypes.SELECT, replacements: { ids } },
-      )
+    queries
+      .getBalances(ids)
       .then(results => sortResults(ids, results, 'CollectiveId'))
       .map(result => get(result, 'balance') || 0),
   );
