+++ server/models/Collective.js
@@ -2136,41 +2136,7 @@ export default function (Sequelize, DataTypes) {
 
   Collective.prototype.getBalance = async function (until) {
     until = until || new Date();
-    const result = await this.sequelize.query(
-      `
-        WITH "blockedFunds" AS (
-          SELECT
-            e."CollectiveId", COALESCE(sum(e.amount), 0) as sum
-          FROM
-            "Expenses" e
-          WHERE
-            e."CollectiveId" = :id
-            AND e."createdAt" < :until
-            AND (
-              e.status = 'SCHEDULED_FOR_PAYMENT'
-              OR (
-                e.status = 'PROCESSING' AND e.data ->> 'payout_batch_id' IS NOT NULL
-              )
-            )
-          GROUP BY
-            e."CollectiveId"
-        )
-
-        SELECT
-          t."CollectiveId",
-          COALESCE(sum(t."netAmountInCollectiveCurrency") - COALESCE(max(bf.sum), 0), 0) AS "balance"
-        FROM
-          "Transactions" t
-        LEFT JOIN "blockedFunds" bf ON t."CollectiveId" = bf."CollectiveId"
-        WHERE
-          t."CollectiveId" = :id
-          AND t."createdAt" < :until
-        GROUP BY
-          t."CollectiveId";
-      `,
-      { type: this.sequelize.QueryTypes.SELECT, replacements: { id: this.id, until } },
-    );
-
+    const result = await queries.getBalances([this.id], until);
     return get(result, '[0].balance') || 0;
   };
